# swagger/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Adventure APIs",
    "description": "API for admin/managing and client/viewing",
    "termsOfService": "http://swagger.io/terms/"
  },
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/friend_request/send_request": {
      "post": {
        "tags": [
          "friend_request"
        ],
        "description": "request add friend",
        "operationId": "request_add_friend",
        "parameters": [
          {
            "name": "token",
            "in": "formData",
            "description": "token",
            "required": true,
            "type": "string"
          },
          {
            "name": "user",
            "in": "formData",
            "description": "user id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "properties": {
                "code": {
                  "type": "integer",
                  "description": "1: success; 0: invalid params; -1: Redis error; -2: DB error; -3: Token is not found; -4 : Can not add yourself; -5 : User id is not existed; -6 : Error. Request sent;  ",
                  "example": 1
                },
                "data": {
                  "type": "object",
                  "description": "in case of success",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "5769097529f87b5c271728e5"
                    },
                    "sender": {
                      "type": "string",
                      "example": "5769097529f87b5c271728e5"
                    },
                    "recipient": {
                      "type": "string",
                      "example": "5769097529f87b5c271728e5"
                    },
                    "status": {
                      "type": "number",
                      "example": 0
                    },
                    "created_at": {
                      "type": "date-time",
                      "example": "2016-10-05T04:27:58.144Z"
                    }
                  }
                },
                "message": {
                  "type": "string",
                  "description": "in case of error",
                  "example": "token is not found"
                }
              }
            }
          }
        }
      }
    },
    "/api/friend_request/confirm": {
      "post": {
        "tags": [
          "friend_request"
        ],
        "description": "confirm add friend",
        "operationId": "confirm_add_friend",
        "parameters": [
          {
            "name": "token",
            "in": "formData",
            "description": "token",
            "required": true,
            "type": "string"
          },
          {
            "name": "_id",
            "in": "formData",
            "description": "friend request id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "properties": {
                "code": {
                  "type": "integer",
                  "description": "1: success; 0: invalid params; -1: Redis error; -2: DB error; -3: Token is not found; -4 : Request is not exits; -5 : Request is not sent; -6 : Can not confirm friend;  ",
                  "example": 1
                },
                "data": {
                  "type": "object",
                  "description": "in case of success",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "5769097529f87b5c271728e5"
                    },
                    "user_one": {
                      "type": "string",
                      "example": "5769097529f87b5c271728e5"
                    },
                    "user_two": {
                      "type": "string",
                      "example": "5769097529f87b5c271728e5"
                    },
                    "created_at": {
                      "type": "date-time",
                      "example": "2016-10-05T04:27:58.144Z"
                    }
                  }
                },
                "message": {
                  "type": "string",
                  "description": "in case of error",
                  "example": "token is not found"
                }
              }
            }
          }
        }
      }
    },
    "/api/friend_request/browse": {
      "get": {
        "tags": [
          "friend_request"
        ],
        "description": "list request friend of an user",
        "operationId": "list_request",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "description": "token",
            "required": true,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "page to pagination",
            "required": false,
            "type": "number"
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "item per page for pagination",
            "required": false,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "properties": {
                "code": {
                  "type": "integer",
                  "description": "1: success; 0: invalid params; -1: Redis error; -2: DB error; -3: Token is not found; -4 : Friend request is not found'; ",
                  "example": 1
                },
                "data": {
                  "type": "object",
                  "description": "in case of success",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "5769097529f87b5c271728e5"
                    },
                    "recipient": {
                      "type": "string",
                      "example": "5769097529f87b5c271728e5"
                    },
                    "sender": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "5769097529f87b5c271728e5"
                        },
                        "first_name": {
                          "type": "string",
                          "example": "Brucelee Thanh"
                        },
                        "last_name": {
                          "type": "string",
                          "example": "Brucelee Thanh"
                        },
                        "password": {
                          "type": "string",
                          "example": "123456"
                        },
                        "email": {
                          "type": "string",
                          "example": "Brucelee Thanh@gmail.com"
                        },
                        "phone_number": {
                          "type": "string",
                          "example": ""
                        },
                        "gender": {
                          "type": "number",
                          "example": 0
                        },
                        "birthday": {
                          "type": "date-time",
                          "example": "1993-07-05T04:27:58.144Z"
                        },
                        "address": {
                          "type": "string",
                          "example": ""
                        },
                        "religion": {
                          "type": "string",
                          "example": ""
                        },
                        "intro": {
                          "type": "string",
                          "example": ""
                        },
                        "fb_id": {
                          "type": "string",
                          "example": ""
                        },
                        "avatar": {
                          "type": "string",
                          "example": ""
                        },
                        "cover": {
                          "type": "string",
                          "example": ""
                        },
                        "created_at": {
                          "type": "date-time",
                          "example": "2016-10-05T04:27:58.144Z"
                        }
                      }
                    },
                    "status": {
                      "type": "number",
                      "example": 0
                    },
                    "created_at": {
                      "type": "date-time",
                      "example": "2016-10-05T04:27:58.144Z"
                    }
                  }
                },
                "total": {
                  "type": "number",
                  "example": 1
                },
                "message": {
                  "type": "string",
                  "description": "in case of error",
                  "example": "token is not found"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/trip_member/request": {
      "post": {
        "tags": [
          "trip_member"
        ],
        "description": "request member trip",
        "operationId": "request",
        "parameters": [
          {
            "name": "token",
            "in": "formData",
            "description": "token",
            "required": true,
            "type": "string"
          },
          {
            "name": "id_trip",
            "in": "formData",
            "description": "id trip",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "properties": {
                "code": {
                  "type": "integer",
                  "description": "1. Redis error; 2. DB error; 3. Token is not found; 4. Trip is not found; 5. This request have existed; 6. You have been invited; 7. You're already member",
                  "example": 1
                },
                "data": {
                  "type": "object",
                  "description": "in case of success",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "5769097529f87b5c271728e5"
                    },
                    "id_trip": {
                      "type": "string",
                      "example": "5769097529f87b5c271728e5"
                    },
                    "owner": {
                      "type": "string",
                      "example": "5769097529f87b5c271728e5"
                    },
                    "created_at": {
                      "type": "date-time",
                      "example": "2016-10-05T04:27:58.144Z"
                    },
                    "status": {
                      "type": "number",
                      "example": 1
                    }
                  }
                },
                "message": {
                  "type": "string",
                  "description": "in case of error",
                  "example": "token is not found"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}